Add support to Rails 5
Also change .travis.yml to run tests with different Rails versions.
