+++ .travis.yml
@@ -1,7 +1,14 @@
 language: ruby
-rvm:
-  - "1.8.7"
-  - "1.9.2"
-  - "1.9.3"
+matrix:
+  include:
+  - rvm: 1.8.7
+    gemfile: test/gemfiles/Gemfile-Rails-3-2
+  - rvm: 1.9.3
+    gemfile: test/gemfiles/Gemfile-Rails-3-2
+  - rvm: 2.3.5
+    gemfile: test/gemfiles/Gemfile-Rails-4-2
+  - rvm: 2.4.2
+    gemfile: test/gemfiles/Gemfile-Rails-5-2
+
 notifications:
   email: false
