+++ test/gemfiles/Gemfile-Rails-3-2
@@ -1,8 +1,10 @@
 source 'https://rubygems.org'
 
+gemspec :path => '../..'
+
 gem 'rake'
 gem 'sqlite3'
-gem 'activerecord', '~> 5.1'
+gem 'activerecord', '~> 3.2.0'
 
 if RUBY_VERSION < '1.9'
   gem 'fastercsv'
