+++ test/gemfiles/Gemfile-Rails-5-1
@@ -1,8 +1,10 @@
 source 'https://rubygems.org'
 
+gemspec :path => '../..'
+
 gem 'rake'
 gem 'sqlite3'
-gem 'activerecord', '~> 5.1'
+gem 'activerecord', '~> 5.1.0'
 
 if RUBY_VERSION < '1.9'
   gem 'fastercsv'
