+++ to_csv_rails.gemspec
@@ -9,7 +9,9 @@ Gem::Specification.new do |s|
   s.homepage    = "http://fernandomarcelo.com/"
   s.summary     = %q{An to_csv implementation for Rails framework}
   s.description = %q{An to_csv implementation focused on making it easier for Rails developers to export data to CSV format.}
-  s.files = Dir['**/*'].select{|f| File.file?(f)}
+  s.files = Dir["lib/**/*", "README.md"]
+
+  s.add_dependency("activerecord", ">= 3.2")
 
   if RUBY_VERSION < "1.9.0"
     s.add_dependency('fastercsv')
